plug in functionality for COTD on welcome page
