+++ src/pages/dashboard/routes/WelcomePage/styles/COTDStyles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+// eslint-disable-next-line max-len
+import { CardsFlexed } from '../../SingleDeck/cardsSection/styles/deckCardStyles';
+
+export const COTDWrapper = styled(CardsFlexed)`
+  width: 100%;
+`;
+
+export const DivWrapper = styled.div``;
